refactor : lottos 의 countRank 부분 개선
- forEach 문으로 Map 의 값을 갱신 해주면서 count 를 했던 부분을
groupingBy 를 사용하여 계산 후, 전체 rank 를 가지고 있는 ranks 와 merge 하는 형식으로 변경

+++ src/main/java/controller/LottoController.java
@@ -67,7 +67,7 @@ public class LottoController {
 		}
 	}
 
-	private LottoResult createLottoResult(Map<Rank, Integer> ranks) {
+	private LottoResult createLottoResult(Map<Rank, Long> ranks) {
 		LottoResult lottoResult = new LottoResult(ranks);
 		OutputView.printRankCounts(lottoResult);
 		return lottoResult;
