+++ src/main/java/domain/LottoResult.java
@@ -5,9 +5,9 @@ import java.util.Map;
 import java.util.Objects;
 
 public class LottoResult {
-	private final Map<Rank, Integer> ranks;
+	private final Map<Rank, Long> ranks;
 
-	public LottoResult(final Map<Rank, Integer> ranks) {
+	public LottoResult(final Map<Rank, Long> ranks) {
 		this.ranks = Objects.requireNonNull(ranks);
 	}
 
@@ -18,7 +18,7 @@ public class LottoResult {
 	private int calculateTotalProfit() {
 		return ranks.entrySet()
 			.stream()
-			.mapToInt(entry -> calculateRankProfit(entry.getKey(), entry.getValue()))
+			.mapToInt(entry -> calculateRankProfit(entry.getKey(), entry.getValue().intValue()))
 			.sum();
 	}
 
@@ -26,7 +26,7 @@ public class LottoResult {
 		return rank.calculateMoney(count);
 	}
 
-	public Map<Rank, Integer> getRanks() {
+	public Map<Rank, Long> getRanks() {
 		return Collections.unmodifiableMap(ranks);
 	}
 }
